bpf: move internal tail-call map to tailcall.h
This is the only part remaining from the maps.h file, move it to a more
appropriate location.
