+++ bpf/bpf_alignchecker.c
@@ -6,7 +6,6 @@
 #include <bpf/config/node.h>
 
 #include "lib/common.h"
-#include "lib/maps.h"
 #include "lib/nat.h"
 #include "lib/trace.h"
 #include "lib/policy_log.h"
