+++ bpf/bpf_host.c
@@ -27,7 +27,6 @@
 #include "lib/config_map.h"
 #include "lib/edt.h"
 #include "lib/arp.h"
-#include "lib/maps.h"
 #include "lib/ipv6.h"
 #include "lib/ipv4.h"
 #include "lib/icmp6.h"
@@ -49,6 +48,7 @@
 #include "lib/host_firewall.h"
 #include "lib/egress_gateway.h"
 #include "lib/srv6.h"
+#include "lib/tailcall.h"
 #include "lib/overloadable.h"
 #include "lib/encrypt.h"
 #include "lib/wireguard.h"
