+++ bpf/tests/drop_notify_test.c
@@ -21,9 +21,9 @@
 #include "bpf/ctx/skb.h"
 #include <bpf/config/node.h>
 
-/* Include lib/maps.h which contains the definition of tail_call_internal first to */
+/* Include lib/tailcall.h which contains the definition of tail_call_internal first to */
 /* avoid it to be included again in lib/drop.h. */
-#include "lib/maps.h"
+#include "lib/tailcall.h"
 
 /* Forward declare the mock func */
 int mock_tail_call(void *ctx, const void *map, __u32 index);
